utf8.c: Remove (mostly) redundant test
The swashes already have the underscore, so this test is redundant.  It
does save some time for this character to avoid having to go out and
load the swash, but why just the underscore?  In fact an earlier commit
changed the macro that most people should use to access this function to
not even call it for the underscore.